requirements.txt
certifi==2017.4.17
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-angular==0.8.4
django-extensions==1.7.9
django-filter==1.0.4
django-invitations==1.9.1
djangorestframework==3.6.3
idna==2.5
Markdown==2.6.8
oauthlib==2.0.2
python3-openid==3.1.0
pytz==2017.2
requests==2.17.3
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
